feat: Homework_2_3_7_8 Updated simulation class
 Updated simulation class

+++ DroneCourseHomework/FILES_CPP/Files/InputFile.cpp
@@ -87,7 +87,7 @@ void InputFile::SetupDronePosition(const std::string& line, DataStructs::InputDa
   }
 
   float x, y, z;
-  for (int index = 0; index < dataAsVector.size(); index++) {
+  for (size_t index = 0; index < dataAsVector.size(); index++) {
     switch (index) {
       case 0:
         SetFloatField(dataAsVector[index], index, &x);
